+++ app/src/main/java/com/tajmoti/tulip/model/DbEpisode.kt
@@ -2,6 +2,7 @@ package com.tajmoti.tulip.model
 
 import androidx.room.Entity
 import androidx.room.ForeignKey
+import com.tajmoti.libtvprovider.show.Episode
 
 @Entity(
     primaryKeys = ["service", "tvShowKey", "seasonKey", "key"],
@@ -17,5 +18,12 @@ data class DbEpisode(
     val tvShowKey: String,
     val seasonKey: String,
     val key: String,
-    val name: String
-)
+    val number: Int?,
+    val name: String?
+) {
+    constructor(service: StreamingService, tvShowKey: String, seasonKey: String, episode: Episode) :
+            this(service, tvShowKey, seasonKey, episode.key, episode.number, episode.name)
+
+    val apiInfo: Episode.Info
+        get() = Episode.Info(key, number, name)
+}