+++ libtvprovider-primewire/src/main/java/com/tajmoti/libprimewiretvprovider/PrimewireEpisodeOrMovie.kt
@@ -1,18 +1,16 @@
 package com.tajmoti.libprimewiretvprovider
 
-import com.tajmoti.libtvprovider.TvItem
-import com.tajmoti.libtvprovider.show.Episode
+import com.tajmoti.libtvprovider.stream.Streamable
 import com.tajmoti.libtvprovider.stream.VideoStreamRef
 import kotlinx.coroutines.*
 import org.jsoup.Jsoup
 import org.jsoup.nodes.Element
 
-data class PrimewireEpisodeOrMovie(
-    override val name: String,
+abstract class PrimewireEpisodeOrMovie(
     private val baseUrl: String,
     internal val episodeUrl: String,
     private val pageLoader: SimplePageSourceLoader
-) : Episode, TvItem.Movie {
+) : Streamable {
     override val key = episodeUrl
 
     override suspend fun loadSources(): Result<List<VideoStreamRef>> {