+++ libtvprovider-primewire/src/main/java/com/tajmoti/libprimewiretvprovider/PrimewireMovie.kt
@@ -1,0 +1,12 @@
+package com.tajmoti.libprimewiretvprovider
+
+import com.tajmoti.libtvprovider.TvItem
+
+class PrimewireMovie(
+    override val name: String,
+    baseUrl: String,
+    episodeUrl: String,
+    pageLoader: SimplePageSourceLoader
+) : PrimewireEpisodeOrMovie(baseUrl, episodeUrl, pageLoader), TvItem.Movie {
+    override val language = "en"
+}