+++ libtvprovider-primewire/src/main/java/com/tajmoti/libprimewiretvprovider/PrimewireSeason.kt
@@ -4,7 +4,7 @@ import com.tajmoti.libtvprovider.show.Season
 
 data class PrimewireSeason(
     override val number: Int,
-    override val episodes: List<PrimewireEpisodeOrMovie>
+    override val episodes: List<PrimewireEpisode>
 ) : Season {
     override val key = number.toString()
 }