+++ libtvprovider/src/main/java/com/tajmoti/libtvprovider/show/Season.kt
@@ -1,13 +1,8 @@
 package com.tajmoti.libtvprovider.show
 
 import com.tajmoti.libtvprovider.Marshallable
-import com.tajmoti.libtvprovider.stream.VideoStreamRef
-import kotlinx.coroutines.async
-import kotlinx.coroutines.awaitAll
-import kotlinx.coroutines.coroutineScope
-import java.io.Serializable
 
-interface Season: Marshallable {
+interface Season : Marshallable {
     /**
      * One-based season number
      */
@@ -15,19 +10,9 @@ interface Season: Marshallable {
 
     val episodes: List<Episode>
 
-
-    suspend fun loadSourcesForAllEpisodes(): List<Pair<Episode, Result<List<VideoStreamRef>>>> {
-        return coroutineScope {
-            val coroutines = episodes.map { async { it to it.loadSources() } }
-            awaitAll(*coroutines.toTypedArray())
-        }
-    }
-
     data class Info(
-        /**
-         * One-based season number
-         */
+        override val key: String,
         val number: Int,
-        val episodeInfo: List<Episode.Info>
-    ): Serializable
+        val episodes: List<Episode.Info>
+    ) : Marshallable
 }