+++ libtvprovider/src/main/java/com/tajmoti/libtvprovider/stream/Streamable.kt
@@ -1,13 +1,8 @@
 package com.tajmoti.libtvprovider.stream
 
 import com.tajmoti.libtvprovider.Marshallable
-import com.tajmoti.libtvprovider.NamedItem
 
-interface Streamable: NamedItem, Marshallable {
+interface Streamable : Marshallable {
 
     suspend fun loadSources(): Result<List<VideoStreamRef>>
-
-    data class Info(
-        val name: String
-    )
 }